A better designed simulation
Both rectification and ideal components are used. Current ripple is really small, the rectified output is filtered with  a 47uf capacitor. There is a 8.1 ohms hand calculated resistor with a 4700uf capacitor (I dont know whether it is that crucial to use a large capacitance there, it could be smaller  since the ripple is allright). The turns ratio is selected as 1 and parameters dor the Lm and leakage etc. are taken from a component found in Digikey for a flyback converter. (592 uH Lm) Yet, there is an increased ampunt of stress on the switch considering the voltage (Uin orders of kVs)

For inductors:
https://www.digikey.com/products/en/transformers/switching-converter-smps-transformers/168?FV=65405b2%2C15c40064%2C15c40066%2C15c40067%2C15c40068%2C15c40069%2C15c4006a%2C15c4006b%2C15c4006c%2C15c4006e%2C15c4006f%2C15c40071%2C15c40072%2C15c40076%2C15c40077%2C15c4000c%2C15c4007b%2C15c40089%2C15c4008e%2C15c40091%2C15c400a5%2C15c400a6%2C15c400af%2C15c400b1%2C15c400b3%2C15c400b6%2C15c400b7%2C15c400b9%2C15c400ba%2C15c400bb%2C15c400bc%2C15c40013%2C15c40015%2C15c40017%2C15c40024%2C15c4002f%2C15c40033%2Cffe000a8&quantity=&ColumnSort=1000011&page=2&k=transformer&pageSize=25&pkeyword=transformer