Design Guideline for all and especially for ripple
Fairchild's very nice guideline.  It has all necessary explanation for design procdure. I have learned that the output ripple we see depends  on the MOSFET drain source peak current, output capacitor ESR, output capacitance value and fsw. Moreover, the snubber resistance affects the MOS peak current. Hence, in order to decrease output ripple we need to increase the Rsnub. (we used 1 kohm, the design recomends 33k). It is good to use a Csnub that can endure the primary side voltage(10 nF-1kV recommended). The multioutput design shows a similar Vout. Here  we see that egp30d is used which is an ultra fastrecovery diode. Also a 10-20 ohm resistance isadded in serieswith this diode to stabilize the light load conditions on the output side.

Additionally, the output is filtered. And the filtering cutoff frequency selection is jsut as before. Hence, we can add an LC filter to decrease the ripple when the capacitor is not capable of handling the high ripple.  This corresponds to a 1/3 of the recommended LC values hence, a 690 nH with 220uF of 69uf to 2.2 uH inductor can meet our needs.

Lastly, the document presents a current control method veeeeery similar to what we have triedin the lab with an optocoupler system. Recommended circuit with  H11A817 Series 4-Pin Phototransistor Optocouplers. Recommended circuit can be implemented.